Add Kubernetes StatefulSet example
This commit adds a basic StatefulSet with a volumeClaimTemplate, so that
a PersistentVolumeClaim and associated PersistentVolume are created for
any resulting pods as desired. Additionally this example includes both
resource limits and securityContext settings, to ensure the appropriate
resources are available, and limit the security boundaries of the
resulting workload.